Create new prop - FusTek_FirePort
Fire Ports are designed for a fire extinguisher nozzle to be inserted
into a panel or an avionics enclosure,, allowing the fire suppressant to
fill the cavity and extinguish the fire.

This prop has been welded to other props using MODEL{} nodes, allowing
easy asset reuse.

(I'll probably need to reposition the Fire Port on the
WhitePanel_Compact prop, to make room for the C&W System prop)